xml-tests: fix test failure due to naive string comparison / refactor
